Moved SphereStore into separate file - Use @Environment to access store
